Move GetBlock to new pattern, add test
