+++ server/service/transport_carves.go
@@ -29,15 +29,3 @@ func decodeCarveBlockRequest(ctx context.Context, r *http.Request) (interface{},
 
 	return req, nil
 }
-
-func decodeGetCarveBlockRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	id, err := idFromRequest(r, "id")
-	if err != nil {
-		return nil, err
-	}
-	blockId, err := idFromRequest(r, "block_id")
-	if err != nil {
-		return nil, err
-	}
-	return getCarveBlockRequest{ID: int64(id), BlockId: int64(blockId)}, nil
-}
